requirements.txt
pre-commit
scikit-image
numpy
paddlepaddle-gpu>=2.2.0
SimpleITK>=2.1.1
PyYAML
pynrrd
tqdm
opencv-python
visualdl
sklearn
nibabel
pydicom
